Refactor bitFlyer adapter into Http/Raw/Adapters/Apis/Facade layers
